refactor(CategoryIcon): 불필요한 CategoryIcon 제거.
