--- src/components/Basic/CategoryIcon.module.css
@@ -1,4 +1,0 @@
-.categoryIcon {
-  width: 36px;
-  height: 36px;
-}
